fix: index activity action filters
