+++ apps/api/src/lib/audit-log.ts
@@ -11,79 +11,17 @@ import { auditLogs } from "@/api/db/schema";
 import type { SafeId } from "@/api/lib/branded-types";
 import { resolveClientIp } from "@/api/lib/client-ip";
 
+import { AUDIT_ACTION, AUDIT_RESOURCE_TYPE } from "./audit-log.constants";
+import type { AuditAction, AuditResourceType } from "./audit-log.constants";
+
+export { AUDIT_ACTION, AUDIT_RESOURCE_TYPE } from "./audit-log.constants";
+export type { AuditAction, AuditResourceType } from "./audit-log.constants";
+
 type ServerLike = {
   requestIP: (request: Request) => { address: string } | null;
 };
 
-export const AUDIT_ACTION = {
-  CREATE: "create",
-  UPDATE: "update",
-  DELETE: "delete",
-  DOWNLOAD: "download",
-  EXECUTE: "execute",
-  ACCESS: "access",
-  CANCEL: "cancel",
-  REVIEW: "review",
-} as const;
-
-export type AuditAction = (typeof AUDIT_ACTION)[keyof typeof AUDIT_ACTION];
-
 export const ORGANIZATION_AUDIT_LOG_RESOURCE_ID = "organization-logs";
-
-export const AUDIT_RESOURCE_TYPE = {
-  AUDIT_LOG: "audit_log",
-  AGENT_SKILL: "agent_skill",
-  AI_MEMORY: "ai_memory",
-  BILLING_CODE: "billing_code",
-  CASE_LAW_MATTER_LINK: "case_law_matter_link",
-  CHAT_FILE: "chat_file",
-  CHAT_MESSAGE: "chat_message",
-  CHAT_THREAD: "chat_thread",
-  CLAUSE: "clause",
-  CLAUSE_CATEGORY: "clause_category",
-  CLAUSE_TEMPLATE_LINK: "clause_template_link",
-  CLAUSE_VARIANT: "clause_variant",
-  CONTACT: "contact",
-  DOCUMENT_TYPE: "document_type",
-  USAGE_ALLOCATION: "usage_allocation",
-  USAGE_ENTITLEMENT: "usage_entitlement",
-  USAGE_EVENT: "usage_event",
-  DESKTOP_EDIT_SESSION: "desktop_edit_session",
-  DOCUMENT_REVIEW_RUN: "document_review_run",
-  ENTITY: "entity",
-  ENTITY_VERSION: "entity_version",
-  EXPENSE: "expense",
-  FIELD: "field",
-  FLOW_DEFINITION: "flow_definition",
-  FLOW_RUN: "flow_run",
-  FOLIO_COLLAB_SESSION: "folio_collab_session",
-  INVOICE: "invoice",
-  MACHINE_API_KEY: "machine_api_key",
-  LEGAL_LIST: "legal_list",
-  LEGAL_LIST_GENERATION: "legal_list_generation",
-  LEGAL_LIST_ITEM: "legal_list_item",
-  MCP_GATEWAY_TOOL: "mcp_gateway_tool",
-  ORGANIZATION_SETTINGS: "organization_settings",
-  PLAYBOOK: "playbook",
-  PROPERTY: "property",
-  RATE_ENTRY: "rate_entry",
-  REPORT_EXPORT: "report_export",
-  RATE_TABLE: "rate_table",
-  SAVED_SEARCH: "saved_search",
-  STYLE_SET: "style_set",
-  TEMPLATE: "template",
-  TIME_ENTRY: "time_entry",
-  USER_FILE: "user_file",
-  VIEW: "view",
-  VIEW_TEMPLATE: "view_template",
-  WORKSPACE: "workspace",
-  WORKSPACE_CONTACT: "workspace_contact",
-  WORKSPACE_MEMBER: "workspace_member",
-  WORK_OBLIGATION: "work_obligation",
-} as const;
-
-export type AuditResourceType =
-  (typeof AUDIT_RESOURCE_TYPE)[keyof typeof AUDIT_RESOURCE_TYPE];
 
 // Generic field-diff shape. Every existing audit payload in the
 // codebase is `{ [field]: { old, new } }` — see entities/move.ts,
